CollTriData.h: replace p_base type cast with uintptr_t

+++ Code/Gel/Collision/CollTriData.h
@@ -106,7 +106,7 @@ private:
 
 		// These two are needed for the cloning function
 		CCollBSPNode *GetBasePointer() const { return (CCollBSPNode*)(m_left_child_and_flags & ~0x3); }
-		void SetBasePointer(CCollBSPNode *p_base) { m_left_child_and_flags = ((uint32) p_base) | (m_left_child_and_flags & 0x3); }
+		void SetBasePointer(CCollBSPNode *p_base) { m_left_child_and_flags = ((uintptr_t) p_base) | (m_left_child_and_flags & 0x3); }
 
 		bool IsLeftGreater() const { return m_left_child_and_flags & mLEFT_IS_GREATER; }
 		void SetLeftGreater(bool greater);
